Changed net() to return a tuple
I needed a way for the caller to get both the response and an
indication that a rate limit was reached.  I couldn't think of a
simpler way to do it, so settled on return a tuple of (response,
exception), even though it feels weird to return an exception object
rather than actually raise an exception.