Solve an occasional crasher and also a matplotlib warning
Running on macos 10.13, i randomly experienced exceptions from
somewhere deep in the pyplot code *sometimes* when imshow was called
on the same file.  The actual error was senseless (i.e., it obviously
didn't indicate the real problem), and only happened when multiple
threads were used.  At first I thought the problem was due to some
race condition or resource use condition during multithreaded
execution, but putting a lock around the imshow() call didn't solve
it.  Finally I rewrote the way that the plot was created (by adding a
subplot), and lo and behold, the crashes no longer seem to happen.
The key seemed to be to create a subplot (and corresponding axes)
object in which the imshow() is called, rather than calling imshow()
on the plt object directly.

The other thing solved in this commit is with "fig, axes =
plt.subplots(...)" it avoids the Matplotlib warning "Adding an axes
using the same arguments as a previous axes currently reuses the
earlier instance.  In a future version, a new instance will always be
created and returned.  Meanwhile, this warning can be suppressed, and
the future behavior ensured, by passing a unique label to each axes
instance."