Catch more error situations
There's some combination of factors that causes occasional very random
errors that seem to be due to bad image file reads or corruption
somewhere or something.  This tries to watch for the symptoms, but I
don't have a good repeatable test case and am not sure this is
enough.  Still, it seems better than nothing.