Fix bug getting amazon credentials
The fact there are two Amazon services but only one AmazonCredentials
object is turning out to be a problem.  AmazonCredentials doesn't know
which of the two services it's being used with.  This is a design
error and will need to be fixed properly.  For now, I'm just avoiding
the problem by creating a fake "amazon" service entry.