Implement support for relaxed comparison
Some of my test cases suggested that some services sometimes did
things like guess the capitalization of a word incorrectly,
particularly when the author's handwriting style used essentially
just larger versions of lower-case letters (e.g., like a bigger "m" at
the beginning of a proper name, or a big "a" instead of A).  These are
understandbly tricky situations for HTR systems.  Another common
problem is that sometimes a service will not guess that a period or
comma was meant to follow immediately after a word, and the service
may put an extra space in front of the punctuation.

There are probably other ways to relax the algorithm.  At these moment
the above are the only cases that I could implement with some
certainty.  More may come to light in the future.