Fix subtle bug causing items in files to be ignored
The loop over the targets at the bottom of targets_from_arguments(...)
was looping over the values of an interator, but also testing the
targets.  This caused the iterator to be consumed the first time
through the loop, and all the rest of the items in the input file to
be ignored.  The fix is to create a list out of the targets before
the loop.