May have solved image overwriting bug
Sometimes, when running with parallel threads enabled (which is the
default), the annotated image of results from two services would be
written on top of each other.  I could not find a fault in my code,
so I started wondering whether there is something happening in a
lower-level function.  Perhaps a buffer is reused, or there's a race
condition, or something.  Provisionally, it appears that what prevents
it is to lock the threads around the call to annotate_image(), which
is the function that calls Matplotlib image and plotting functions.
The changes in this file are around the function _send(...).