Require boto3 to be version 1.10 or higher
I was getting errors like this:

amazon-rekognition failed: Error: 'endpoint_resolver'

They were infrequent, but sometimes they would start happening every
time I ran Handprint.  It appears to be due to multithreading issues.
Since updating to a more recent version of boto3, I haven't seen it
happen, so I hope this solves it.  But even if it doesn't, it's still
a good idea to move on to the 1.10 series of boto3.