Try shutils.copy() if shutils.copy2() fails
For some reason, on my macOS 10.13 vm in Parallels, copy2 fails, but
copy succeeds. Seems to be some sort of permissions issue.  Maybe
or writing the metadata is problematic, and maybe the problem is
related to the source rather than the destination, but in any case,
failling back to copy() is working.